+++ craft_cli/completion/__init__.py
@@ -1,0 +1,8 @@
+"""Bash completion script generation module."""
+
+from .completion import complete
+
+
+__all__ = [
+    "complete",
+]